发布最新功能到release

+++ release/deploy/deploy-docker-deps.sh
@@ -2,6 +2,8 @@
 
 PACKAGE_PATH=${PACKAGE_PATH:-$HOME/dashboard_packages}
 DOCKER_VERSION=${DOCKER_VERSION:-"1.9.1"}
+
+source ./deploy-check-deps.sh
   
 function get_docker_deps_deb() {
   echo "Package path: $PACKAGE_PATH"
@@ -77,7 +79,7 @@ function get_docker_deps_deb() {
     fi
   fi
   echo " ... done"
-  
+ 
   echo -n "get libsystemd-journal0"
   if [ ! -f ${PACKAGE_PATH}/docker/libsystemd-journal0_204-5ubuntu20.19_amd64.deb ]; then
     wget http://launchpadlibrarian.net/242858416/libsystemd-journal0_204-5ubuntu20.19_amd64.deb -P ${PACKAGE_PATH}/docker >& /dev/null
@@ -105,9 +107,24 @@ function get_docker_deps_deb() {
 
 function install_docker_deps_apt() {
   sudo apt-get update
-  sudo apt-get -y install apt-transport-https ca-certificates
+  result=`check_deps apt-transport-https`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install apt-transport-https >& /dev/null
+  fi
 
-  sudo apt-get -y install linux-image-extra-$(uname -r) linux-image-extra-virtual
+  result=`check_deps ca-certificates`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install ca-certificates >& /dev/null
+  fi
+
+  result=`check_deps linux-image-extra-$(uname -r)`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install linux-image-extra-$(uname -r) >& /dev/null
+  fi
+  result=`check_deps linux-image-extra-virtual`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install linux-image-extra-virtual >& /dev/null
+  fi
 
   sudo apt-key adv --keyserver hkp://p80.pool.sks-keyservers.net:80 --recv-keys 58118E89F3A912897C070ADBF76221572C52609D
 
@@ -115,34 +132,46 @@ function install_docker_deps_apt() {
 
   sudo apt-get update
 
-  sudo apt-get purge lxc-docker
+  result=`check_deps lxc-docker`
+  if [ $result == "no" ]; then
+    sudo apt-get purge lxc-docker
+  fi
 
   sudo apt-cache policy docker-engine
 
 
-  sudo apt-get -y install docker-engine
+  result=`check_deps docker-engine`
+  if [ $result == "no" ]; then
+    sudo apt-get -y install docker-engine >& /dev/null
+  fi
 }
 
 function install_docker_deps_dpkg() {
-  echo -n "Install apt-transport-https"
-  sudo dpkg -i ${PACKAGE_PATH}/docker/apt-transport-https_1.0.1ubuntu2.14_amd64.deb >& /dev/null
-  if [ $? -ne 0 ]; then
-    echo " ... failed"
-    echo " please find another resource for the package - apt-transport-https_1.0.1ubuntu2.14_amd64.deb"
-    echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
-    exit 125
+  result=`check_deps apt-transport-https`
+  if [ $result == "no" ]; then
+    echo -n "Install apt-transport-https"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/apt-transport-https_1.0.1ubuntu2.14_amd64.deb >& /dev/null
+    if [ $? -ne 0 ]; then
+      echo " ... failed"
+      echo " please find another resource for the package - apt-transport-https_1.0.1ubuntu2.14_amd64.deb"
+      echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
+      exit 125
+    fi
+    echo " ... done"
   fi
-  echo " ... done"
 
-  echo -n "Install ca-certificates"
-  sudo dpkg -i ${PACKAGE_PATH}/docker/ca-certificates_20160104ubuntu0.14.04.1_all.deb >& /dev/null
-  if [ $? -ne 0 ]; then
-    echo " ... failed"
-    echo " please find another resource for the package - ca-certificates_20160104ubuntu0.14.04.1_all.deb"
-    echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
-    exit 125
+  result=`check_deps ca-certificates`
+  if [ $result == "no" ]; then
+    echo -n "Install ca-certificates"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/ca-certificates_20160104ubuntu0.14.04.1_all.deb >& /dev/null
+    if [ $? -ne 0 ]; then
+      echo " ... failed"
+      echo " please find another resource for the package - ca-certificates_20160104ubuntu0.14.04.1_all.deb"
+      echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
+      exit 125
+    fi
+    echo " ... done"
   fi
-  echo " ... done"
   
   #sudo dpkg -i ${PACKAGE_PATH}/linux-image-extra-$(uname -r).deb
   #sudo dpkg -i ${PACKAGE_PATH}/linux-image-extra-virtual_3.13.0.93.100_amd64.deb
@@ -150,21 +179,42 @@ function install_docker_deps_dpkg() {
   echo "Purge the old repo if it exists."
   sudo apt-get purge lxc-docker >& /dev/null
   
-  echo "Install extra packages:  aufs-tools cgroup-lite libltdl7 libsystemd-journal0"
-  sudo dpkg -i ${PACKAGE_PATH}/docker/aufs-tools_3.2+20130722-1.1_amd64.deb >& /dev/null
-  sudo dpkg -i ${PACKAGE_PATH}/docker/cgroup-lite_1.9_all.deb >& /dev/null
-  sudo dpkg -i ${PACKAGE_PATH}/docker/libltdl7_2.4.2-1.7ubuntu1_amd64.deb >& /dev/null
-  sudo dpkg -i ${PACKAGE_PATH}/docker/libsystemd-journal0_204-5ubuntu20.19_amd64.deb >& /dev/null
-  
-  echo -n "Install Docker"
-  sudo dpkg -i ${PACKAGE_PATH}/docker/docker-engine_${DOCKER_VERSION}-0~trusty_amd64.deb >& /dev/null
-  if [ $? -ne 0 ]; then
-    echo " ... failed"
-    echo " please find another resource for the package - docker-engine_${DOCKER_VERSION}-0~trusty_amd64.deb"
-    echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
-    exit 125
+  result=`check_deps aufs-tools`
+  if [ $result == "no" ]; then
+    echo "Install extra packages:  aufs-tools"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/aufs-tools_3.2+20130722-1.1_amd64.deb >& /dev/null
   fi
-  echo " ... done"
+
+  result=`check_deps cgroup-lite`
+  if [ $result == "no" ]; then
+    echo "Install extra packages: cgroup-lite"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/cgroup-lite_1.9_all.deb >& /dev/null
+  fi
+
+  result=`check_deps libltdl7`
+  if [ $result == "no" ]; then
+    echo "Install extra packages: libltdl7"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/libltdl7_2.4.2-1.7ubuntu1_amd64.deb >& /dev/null
+  fi
+ 
+  result=`check_deps libsystemd-journal0`
+  if [ $result == "no" ]; then
+    echo "Install extra packages: libsystemd-journal0"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/libsystemd-journal0_204-5ubuntu20.19_amd64.deb >& /dev/null
+  fi
+  
+  result=`check_deps docker-engine`
+  if [ $result == "no" ]; then
+    echo -n "Install Docker"
+    sudo dpkg -i ${PACKAGE_PATH}/docker/docker-engine_${DOCKER_VERSION}-0~trusty_amd64.deb >& /dev/null
+    if [ $? -ne 0 ]; then
+      echo " ... failed"
+      echo " please find another resource for the package - docker-engine_${DOCKER_VERSION}-0~trusty_amd64.deb"
+      echo " download it and put it to the path: ${PACKAGE_PATH}/docker"
+      exit 125
+    fi
+    echo " ... done"
+  fi
 }
 
 function uninstall_docker_deps() {
